code/requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=3.0.0
tqdm>=4.66.0
